x86/dom0: only disable SMAP for the PV dom0 build
The PVH dom0 builder doesn't switch page tables and has no need to run with
SMAP disabled.

Use stac() and clac(), as that's safe to use even if events would hit in the
middle of the region with SMAP disabled.  Nesting stac() and clac() calls
is not safe, but it's not the case here, as the domain builder explicitly uses
raw_{copy_to,clear}_guest() accessors which don't toggle SMAP.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>