x86/mm: ensure L4 idle_pg_table is not modified past AP bringup
The idle_pg_table L4 is cloned to create all the other L4 Xen uses, and hence
it shouldn't be modified once further L4 are created.

Further changes in this series will add a per-CPU idle page tables, and hence
we need to ensure that by the time APs are started the BSP L4 idle page
directory is not changed, as otherwise the copies in the APs would get out of
sync.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
Changes since v1:
 - Expand commit message to reason about the SYS_STATE_smp_boot check.
 - Reword comment message to not mention dom0 build explicitly, as it would
   become stale by further changes.