x86/mm: introduce support to populate a per-CPU page-table region
Add logic in map_pages_to_xen() and modify_xen_mappings() so that TLB flushes
are only performed locally when dealing with entries in the per-CPU area of the
page-tables.

No functional change intended, as there are no callers added that setup of
modify per-CPU mappings, nor is the per-CPU area still properly setup in
the page-tables yet.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>