x86/pv: allow using a unique a per-pCPU root page table
When running PV guests it's possible for the guest to use the same root page
table (L4) for all vCPUs, which in turn will result in Xen also using the same
root page table on all pPCU that are running any domain vCPU.

With XPTI Xen switches to a per-CPU shadow L4 when running in guest context,
switching to the fully populated L4 when in Xen context.

Take advantage of this existing shadowing and force the usage of a per-CPU L4
that shadows the guest selected L4 when Address Space Isolation is requested
for PV guests.

In order to map the guest L4 in a per-CPU slot the CPU needs to be using a
per-CPU L4.  Account for this and only attempt to map the guest L4 once the CPU
has already loaded the per-CPU shadow L4.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>