x86/mm: switch to a per-CPU mapped stack when using ASI
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
---
There's further work required in order to allocate the stack from domheap
instead of xenheap, otherwise the per-CPU mapping is not really helpful.

Should the stack always be mapped in the per-CPU VA range regardless of whether
ASI is active on the system for any domain?  That might simplify some of the
logic.