x86/smp: remove call to fixup_irqs() on shutdown
It's my understanding the purpose of the fixup_irqs() call in smp_send_stop()
is to move any interrupts still active to the CPU coordinating the shutdown
(and the one from which kexec will be handed over).

However the current code in fixup_irqs() doesn't behave like that.  The
input mask parameter signals from which CPUs interrupts should be evacuated,
but not to which set of CPUs those interrupts should be assigned.  The logic in
fixup_irqs() will evacuate interrupts from any CPUs not on the input mask, but
the destination of those interrupts will be any CPU marked as online.  Since
the call to fixup_irqs() happens ahead of marking CPUs as offline, this results
in pointless shuffling of interrupts around.

Even if fixup_irqs() did the expected thing and attempted to move all
interrupts to the input CPU mask, there's no guarantee that all sources will be
in a state that allows moving them.  Some IRQ descriptors might have
move_cleanup_count or move_in_progress set, thus preventing the move.  It's
also possible the destination CPU might not have enough vectors to accommodate
all active interrupts.

Seeing as the call to fixup_irqs() in smp_send_stop() is not fulfilling it's
intended purpose, remove it.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>