x86/pv: gate setting per-domain slot to XPTI being active for the domain
It's possible to have XPTI not active for all PV domains (active for domUs,
inactive for dom0), hence don't gate setting the per-domain slot on the
presence of the per-pCPU shadow root page-table.  Instead set the slot based on
whether XPTI is active for the domain.  This avoid pointlessly setting the
per-domain slot if the shadow root page-table won't be used by the domain.

Fixes: 0d3e7f0b6bf0 ('xen/x86: support per-domain flag for xpti')
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>