x86/mm: replace {create,destroy}_perdomain_mapping() implementations to use map_pages()
The current logic in {create,destroy}_perdomain_mapping() is an ad-hoc
implementation that modifies page-tables to fulfill the usages required by the
per-domain slot.

Remove such open-coded page-table manipulations and instead delegate page-table
changes to map_pages() and destroy_mappings().  Since such functions require a
root page table as parameter, use a scratch L4 only populated with the
per-domain slot as the parameter.

If the caller has requested the region to be populated, such filling is done by
fetching the L1 page-table after map_pages() has created the paging structures
and populating the slots directly.  Otherwise each entry to be populated would
require a call to map_pages() which is wasteful when we know the L1 table is
unconditionally present.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>