x86/mm: set Xen GDT L1 entry using map_pages()
The current code to update the Xen part of the GDT when running a PV guest
relies on caching the direct map address of all the L1 tables used to map the
GDT and LDT, so that entries can be modified.

In and attempt to remove those references, use the newly introduced map_pages()
with a guest root page-table pointer so setup the Xen mappings for the ot be
loaded guest page-table.

Note that at the point update_xen_slot_in_full_gdt() gets called the root
page-table of v is not yet the loaded once on the pCPU, hence the recursive
linear mappings cannot be used to access the L1 entry in a faster way.  This is
however only used when context switching vCPUs.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>