x86/mm: don't flush TLB when modifying non-idle page-tables
The Xen page-table manipulation functions always flush the TLB, as it assumes
the modified page-table is the one currently in-use.

Limit the TLB flushing to modifications done to the idle page-table, when
modifying a different page table leave the flushing to the caller if required.

No functional change intended, as no caller passes a root page-table different
than the idle one.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>