x86/setup: fix bootstrap_map_addr() usage of destroy_xen_mappings()
bootstrap_map_addr() clearly states that it doesn't indented to remove the L2
tables, as it should be re-used to create new mappings.  To that end, it should
use destroy_xen_mappings() which will free empty L2 tables.

Fix this by using map_pages_to_xen(), which does zap the page-table entries,
but does not free page-table structures.

Fixes: 4376c05c3113 ('x86-64: use 1GB pages in 1:1 mapping if available')
Signed-off-by: Roger Pau Monné <roger.pau@ctrix.com>
---
The fixes tag reflects the fact that if 4376c05c3113 freed the L2 correctly
when empty, it would have become obvious that bootstrap_map_addr() shouldn't be
using it if it wants to keep the L2.