x86/mm: special case super page alignment detection for INVALID_MFN
INVALID_MFN is ~0, so it doesn't pass the alignment check to be used by super
page slots (either in L2 or L3).  Special case INVALID_MFN so it's considered
to be aligned for all slots.

This fixes a regression introduced by 0b6b51a69f4d, where the switch from 0 to
INVALID_MFN caused all super-pages to be shattered when attempting to remove
mappings by passing INVALID_MFN instead of 0.

Fixes: 0b6b51a69f4d ('xen/mm: Switch map_pages_to_xen to use MFN typesafe')
Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>