x86/mm: limit non-idle page-table manipulations to the per-domain area
Page-tables used by Xen, either the idle page-table, the HVM monitor ones, or
the page-tables built by PV guests all share a set of slots that are owned by
Xen.  Limit which modifications can be made to non-idle page-tables and only
allow modifying the per-domain slot region, which is different and set based on
the domain owning the page-table.

With the added limitation of which area can be manipulated on non-idle
page-tables also lessen the locking and TBL flushing discipline.  When
modifying the per-domain area the caller is responsible of ensuring correctness
of concurrent accesses; either by external locking, or by not concurrently
modifying the same entries.  Similarly the caller is also responsible for doing
any TLB flushing if necessary.

flush_area() has now a dependency on 'root_pgt' being declared in the caller
context, and containing the address of the root page-table being modified.
Note the scope of flush_area() is limited to the functions where it's used, so
the dependency is better than the churn introduced by modifying every caller to
pass 'root_pgt' as a macro parameter.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>